:speak flow key
Added support for a :speak function key in a flow block definition. This means if the webserver is up, the output of the :speak key can be spoken via ElevenLabs text-to-speech API. Use the settings panel to add your API key and choose the voice model. (all optional of course, but can be fun for notifications and adding another layer of "observability").